% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjsc.R
\name{adjsc}
\alias{adjsc}
\title{Spectral clustering to the adjacency matrix}
\usage{
adjsc(adj, nclust, kmeans_iter = 1000, kmeans_rep = 100)
}
\description{
For details, see paper "CONSISTENCY OF SPECTRAL CLUSTERING IN STOCHASTIC BLOCK MODELS" by Jing Lei and Alessandro Rinaldo
}
