% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bfclust_addmin}
\alias{bfclust_addmin}
\title{SDP-bf adding an entry upper bound constraint}
\usage{
bfclust_addmin(S, mmin, nclust = as.integer(c()), maxiter = 100L,
  tolerance = 0.01, admm_penalty = 100, verbose = 0L)
}
\arguments{
\item{S}{Input symmetric matrix}

\item{mmin}{Minimum cluster size}

\item{nclust}{Vector of candidate numbers of clusters}

\item{maxiter}{Max number of iteration}

\item{tolerance}{ADMM tolerance}

\item{admm_penalty}{ADMM penalty parameter}

\item{verbose}{Level of verbosity}
}
\value{
A list containing SDP results
}
\description{
SDP-bf adding an entry upper bound constraint
}
