% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsc.R
\name{rsc}
\alias{rsc}
\title{Regularized spectral clustering}
\usage{
rsc(adj, nclust, const = 0.25, kmeans_iter = 1000, kmeans_rep = 100)
}
\description{
Regularized spectral clustering where entries of the
adjacency matrix are inflated by a constant small number. 
See paper "PSEUDO-LIKELIHOOD METHODS FOR COMMUNITY DETECTION IN LARGE SPARSE NETWORKS" by ARASH A. AMINI, AIYOU CHEN, PETER J. BICKEL AND ELIZAVETA LEVINA for details"
}
