% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sdp_caili}
\alias{sdp_caili}
\title{SDP-CL method}
\usage{
sdp_caili(S, lambda = as.numeric(c()), maxiter = 100L, tolerance = 0.01,
  admm_penalty = 1, verbose = 0L)
}
\arguments{
\item{S}{Input symmetric matrix}

\item{lambda}{Vector of possible values of SDP tuning parameter}

\item{maxiter}{Max number of iteration}

\item{tolerance}{ADMM tolerance}

\item{admm_penalty}{ADMM penalty parameter}

\item{verbose}{Level of verbosity}
}
\value{
A list containing SDP results
}
\description{
This function computes a solution path of SDP-CL 
under a sequence of values of tuning parameter lambda
(see ROBUST AND COMPUTATIONALLY FEASIBLE COMMUNITY DETECTION IN THE PRESENCE OF ARBITRARY OUTLIER NODES by T. Tony Cai and Xiaodong Li)
}
