% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sdp_gv}
\alias{sdp_gv}
\title{SDP-GV method}
\usage{
sdp_gv(S, lambda = as.numeric(c()), maxiter = 1000000L, tolerance = 0.01,
  admm_penalty = 100, verbose = 0L)
}
\arguments{
\item{S}{Input symmetric matrix}

\item{lambda}{Vector of possible values of SDP tuning parameter}

\item{maxiter}{Max number of iteration}

\item{tolerance}{ADMM tolerance}

\item{admm_penalty}{ADMM penalty parameter}

\item{verbose}{Level of verbosity}
}
\value{
A list containing SDP results
}
\description{
This function computes a solution path of SDP-GV
under a sequence of values of tuning paramter lambda
(see COMMUNITY DETECTION IN SPARSE NETWORKS VIA GROTHENDIECK’S INEQUALITY by OLIVIER GUE ́DON AND ROMAN VERSHYNIN)
}

